+++ NEWS.md
@@ -16,7 +16,9 @@ Major Enhancements
 
 Minor Enhancements
 
-## 4.0.0
+## 4.0.2
+
+Released 2022-05-18
 
 Rewrite by https://github.com/aartiPl